requirements.txt
absl-py
astunparse
cachetools
certifi
chardet
click
Flask
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
itsdangerous
Jinja2
Keras-Preprocessing
Markdown
MarkupSafe
numpy
oauthlib
opt-einsum
Pillow
protobuf
pyasn1
pyasn1-modules
requests
requests-oauthlib
rsa
six
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
typing-extensions
urllib3
Werkzeug
wrapt